third_party/blink/web_tests/fast/events/init-custom-event-isolated-world-expected.txt

CONSOLE MESSAGE: CustomEvent received in main world
CONSOLE MESSAGE: detail was {"foo":5,"bar":"hello","targetWorld":"main"}
CONSOLE MESSAGE: CustomEvent received in isolated world
CONSOLE MESSAGE: detail was {"foo":5,"bar":"hello","targetWorld":"isolated"}
Tests that properties of CustomEvent initialized with initCustomEvent() are cloned when accessed in isolated worlds.
